Function quill_borrow_check::borrow_check

pub fn borrow_check(
    source_file: &SourceFileIdentifier, 
    mir: SourceFileMIR
) -> DiagnosticResult<SourceFileMIR>

Checks to make sure that borrows of data do not outlive the data they borrow, and to make sure that once data is not used when it is not owned. This mostly implements the algorithm described here, notably excluding mutable references, since Quill does not have a notion of mutable references.